+++ packages/apple_vision_face_mesh/lib/apple_vision_face_mesh.dart
@@ -1,0 +1,5 @@
+library apple_vision_face_mesh;
+
+export 'src/face_mesh_controller.dart';
+export 'src/face_mesh_data.dart';
+