+++ packages/apple_vision_image_depth/lib/src/image_depth_controller.dart
@@ -4,6 +4,30 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:apple_vision_commons/apple_vision_commons.dart';
 
+class ImageDepthData{
+  /// Process the image using apple vision and return the requested information or null value
+  /// 
+  /// [image] as Uint8List is the image that needs to be processed
+  /// this needs to be in an image format raw will not work.
+  /// 
+  /// [imageSize] as Size is the size of the image that is being processed
+  /// 
+  /// [format] the output format of the image
+  ImageDepthData({
+    required this.image,
+    required this.imageSize,
+    this.format = PictureFormat.png,
+    this.orientation = ImageOrientation.up,
+    this.confidence = 0.75
+  });
+  /// Image to be processed
+  Uint8List image;
+  Size imageSize;
+  double confidence;
+  PictureFormat format;
+  ImageOrientation orientation;
+}
+
 /// The [AppleVisionImageDepthController] holds all the logic of this plugin,
 /// where as the [AppleVisionObject] class is the frontend of this plugin.
 class AppleVisionImageDepthController {
@@ -17,18 +41,19 @@ class AppleVisionImageDepthController {
   /// [imageSize] as Size is the size of the image that is being processed
   /// 
   /// [orientation] The orientation of the image
-  Future<Uint8List?> processImage(Uint8List image, Size imageSize, [double confidence = 0.75,ImageOrientation orientation = ImageOrientation.up]) async{
+  Future<Uint8List?> processImage(ImageDepthData data) async{
     try {
-      final data = await _methodChannel.invokeMapMethod<String, dynamic>(  
+      final result = await _methodChannel.invokeMapMethod<String, dynamic>(  
         'process',
-        {'image':image,
-          'width': imageSize.width,
-          'height':imageSize.height,
-          'confidence': confidence,
-          'orientation': orientation.name
+        {'image':data.image,
+          'width': data.imageSize.width,
+          'height':data.imageSize.height,
+          'confidence': data.confidence,
+          'orientation': data.orientation.name,
+          'format': data.format.name,
         }
       );
-      return _convertData(data);
+      return _convertData(result);
     } catch (e) {
       debugPrint('$e');
     }